Bridge classification between portfolio's system and 2dii's standard

This function adds two columns, sector and borderline to the input portfolio, corresponding to the bridged sector classification.

bridge_sector(data)

Arguments

data
A loanbook dataframe.

Value

A loanbook dataframe with additional sector and borderline columns.

Details

The r2dii.dataraw package must be attached (i.e. run library(r2dii.dataraw) before you call bridge_sector()).

See also

Examples

# Must be attached
library(r2dii.dataraw)

bridge_sector(r2dii.dataraw::loanbook_demo)