Match a loanbook and asset-level datasets by the name_* columns

match_name() scores the match between names in a loanbook dataset (columns name_direct_loantaker and name_ultimate_parent) with names in an asset-level dataset (colum n name_company). The raw names are first transformed and stored in a simple_name column, then the similarity between the simple_name columns in each of the loanbook and ald datasets is scored using stringdist::stringsim().

match_name(
  loanbook,
  ald,
  by_sector = TRUE,
  min_score = 0.8,
  method = "jw",
  p = 0.1,
  overwrite = NULL
)

Arguments

loanbook
Dataframes with simpler_name and optionally sector columns.

ald
Dataframes with simpler_name and optionally sector columns.

by_sector
Should the combinations be done by sector?

min_score
A number (length-1) to set the minimum score values you want to pick.

method
Method for distance calculation. The default is "osa", see stringdist-metrics.

p
Penalty factor for Jaro-Winkler distance. The valid range for p is 0 <= p <= 0.25. If p=0 (default), the Jaro-distance is returned. Applies only to method='jw'.

overwrite
A dataframe used to overwrite the sector and/or name columns of a particular direct loantaker or ultimate parent. To overwrite only sector, the value in the name column should be NA.

Value

A dataframe with the same columns as the loanbook data with additional columns: id, sector_x, source, simpler_name_x, simpler_name_y, score, name_y, sector_y.

Details

The process to create the simple_name columns applies best practices commonly used in name matching algorithms, such as:

Remove special characters.

Replace language specific characters.

Abbreviate certain names to reduce their importance in the matching.

Spell out numbers to increase their importance.

See also

Examples

# Use tibble()
library(dplyr)
library(r2dii.dataraw)

match_name(loanbook_demo, ald_demo)

#> # A tibble: 833 x 27
#>    simpler_name_x simpler_name_y score id    sector_x source name_y sector_y
#>    <chr>          <chr>          <dbl> <chr> <chr>    <chr>  <chr>  <chr>   
#>  1 astonmartin    astonmartin    1     UP23  automot… loanb… aston… automot…
#>  2 avtozaz        avtozaz        1     UP25  automot… loanb… avtoz… automot…
#>  3 bogdan         bogdan         1     UP36  automot… loanb… bogdan automot…
#>  4 chauto         chauto         1     UP52  automot… loanb… ch au… automot…
#>  5 chauto         chtcauto       0.867 UP52  automot… loanb… chtc … automot…
#>  6 chehejia       chehejia       1     UP53  automot… loanb… chehe… automot…
#>  7 chtcauto       chauto         0.867 UP58  automot… loanb… ch au… automot…
#>  8 chtcauto       chtcauto       1     UP58  automot… loanb… chtc … automot…
#>  9 dongfenghonda  dongfenghonda  1     UP79  automot… loanb… dongf… automot…
#> 10 dongfenghonda  dongfengluxgen 0.867 UP79  automot… loanb… dongf… automot…
#> # … with 823 more rows, and 19 more variables: name_ultimate_parent <chr>,
#> #   name_direct_loantaker <chr>, id_loan <chr>, id_direct_loantaker <chr>,
#> #   id_intermediate_parent_1 <chr>, name_intermediate_parent_1 <chr>,
#> #   id_ultimate_parent <chr>, loan_size_outstanding <dbl>,
#> #   loan_size_outstanding_currency <chr>, loan_size_credit_limit <dbl>,
#> #   loan_size_credit_limit_currency <chr>, sector_classification_system <chr>,
#> #   sector_classification_input_type <chr>,
#> #   sector_classification_direct_loantaker <dbl>, fi_type <chr>,
#> #   flag_project_finance_loan <chr>, name_project <lgl>,
#> #   lei_direct_loantaker <lgl>, isin_direct_loantaker <lgl>

match_name(
  loanbook_demo, ald_demo,
  min_score = 0.5,
  by_sector = FALSE
)

#> # A tibble: 174,575 x 27
#>    simpler_name_x simpler_name_y score id    sector_x source name_y sector_y
#>    <chr>          <chr>          <dbl> <chr> <chr>    <chr>  <chr>  <chr>   
#>  1 abahydropower… abahydropower… 1     UP1   power    loanb… aba h… power   
#>  2 abahydropower… achinskyglino… 0.535 UP1   power    loanb… achin… cement  
#>  3 abahydropower… affinityrenew… 0.598 UP1   power    loanb… affin… power   
#>  4 abahydropower… africaoil corp 0.599 UP1   power    loanb… afric… oil&gas 
#>  5 abahydropower… africonshp sa  0.581 UP1   power    loanb… afric… shipping
#>  6 abahydropower… agnisteelspri… 0.535 UP1   power    loanb… agni … power   
#>  7 abahydropower… agrenewables   0.615 UP1   power    loanb… agren… power   
#>  8 abahydropower… airasiaxbhd    0.545 UP1   power    loanb… airas… aviation
#>  9 abahydropower… airbaltic      0.545 UP1   power    loanb… airba… aviation
#> 10 abahydropower… airblue        0.551 UP1   power    loanb… airbl… aviation
#> # … with 174,565 more rows, and 19 more variables: name_ultimate_parent <chr>,
#> #   name_direct_loantaker <chr>, id_loan <chr>, id_direct_loantaker <chr>,
#> #   id_intermediate_parent_1 <chr>, name_intermediate_parent_1 <chr>,
#> #   id_ultimate_parent <chr>, loan_size_outstanding <dbl>,
#> #   loan_size_outstanding_currency <chr>, loan_size_credit_limit <dbl>,
#> #   loan_size_credit_limit_currency <chr>, sector_classification_system <chr>,
#> #   sector_classification_input_type <chr>,
#> #   sector_classification_direct_loantaker <dbl>, fi_type <chr>,
#> #   flag_project_finance_loan <chr>, name_project <lgl>,
#> #   lei_direct_loantaker <lgl>, isin_direct_loantaker <lgl>