Pick rows where score is 1 and level per loan is of highest priority

Pick rows where score is 1 and level per loan is of highest priority

prioritize(data, priority = NULL)

Arguments

data
A dataframe, commonly the output of match_name().

priority
One of:

NULL: defaults to the default level priority as returned by prioritize_level().

A character vector giving a custom priority.

A function to apply to the output of prioritize_level(), e.g. rev.

A quosure-style lambda function, e.g. ~ rev(.x).

Value

A dataframe with a single row per loan, where score is 1 and priority level is highest.

Handling grouped data

This function ignores but preserves existing groups them.

See also

Examples

library(dplyr)

# styler: off
matched <- tribble(
  ~sector, ~sector_ald,  ~score, ~id,  ~level,
  "coal",  "coal",       1,      "aa", "ultimate_parent",
  "coal",  "coal",       1,      "aa", "direct_loantaker",
  "coal",  "coal",       1,      "bb", "intermediate_parent",
  "coal",  "coal",       1,      "bb", "ultimate_parent",
)
# styler: on

prioritize_level(matched)

#> [1] "direct_loantaker"    "intermediate_parent" "ultimate_parent"    

# Using default priority
prioritize(matched)

#> # A tibble: 2 x 5
#>   sector sector_ald score id    level              
#>   <chr>  <chr>      <dbl> <chr> <chr>              
#> 1 coal   coal           1 aa    direct_loantaker   
#> 2 coal   coal           1 bb    intermediate_parent

# Using the reverse of the default priority
prioritize(matched, priority = rev)

#> # A tibble: 2 x 5
#>   sector sector_ald score id    level          
#>   <chr>  <chr>      <dbl> <chr> <chr>          
#> 1 coal   coal           1 aa    ultimate_parent
#> 2 coal   coal           1 bb    ultimate_parent

# Same
prioritize(matched, priority = ~ rev(.x))

#> # A tibble: 2 x 5
#>   sector sector_ald score id    level          
#>   <chr>  <chr>      <dbl> <chr> <chr>          
#> 1 coal   coal           1 aa    ultimate_parent
#> 2 coal   coal           1 bb    ultimate_parent

# Using a custom priority
bad_idea <- c("intermediate_parent", "ultimate_parent", "direct_loantaker")

prioritize(matched, priority = bad_idea)

#> # A tibble: 2 x 5
#>   sector sector_ald score id    level              
#>   <chr>  <chr>      <dbl> <chr> <chr>              
#> 1 coal   coal           1 bb    intermediate_parent
#> 2 coal   coal           1 aa    ultimate_parent