r2dii.match: Tools to Match Financial Portfolios with Climate Data

These tools implement in R a fundamental part of the software 'PACTA' (Paris Agreement Capital Transition Assessment), which is a free tool that calculates the alignment between financial portfolios and climate scenarios (<https://2degrees-investing.org/>). Financial institutions use 'PACTA' to study how their capital allocation impacts the climate. This package matches data from financial portfolios to asset level data from market-intelligence databases (e.g. power plant capacities, emission factors, etc.). This is the first step to assess if a financial portfolio aligns with climate goals.

See also

Author

Maintainer: Mauro Lepore maurolepore@gmail.com (ORCID) [contractor]

Authors:

Jackson Hoffart jackson@2degrees-investing.org (ORCID)

Klaus Hagedorn klaus@2degrees-investing.org

Florence Palandri florence@2degrees-investing.org

Evgeny Petrovsky

Other contributors:

2 Degrees Investing Initiative contact@2degrees-investing.org [copyright holder, funder]