Replace customer names

replace_customer_name() takes any character vector, usually a vector of customer names, and replaces (a) to lower case; (b) to latin-ascii characters; and (c) to standard abbreviations of ownership types.

get_replacements() allows you to access the default replacements table, so you can amend it and pass it to replace_customer_name() via the from_to argument.

replace_customer_name(
  x,
  from_to = NULL,
  ownership = NULL,
  remove_ownership = FALSE
)

get_replacements()

Arguments

x
Character string, (likely) containing customer names.

from_to
A dataframe with replacement rules to be applied, contains columns from (for initial values) and to (for resulting values).

ownership
vector of company ownership types to be distinguished for cut-off or separation.

remove_ownership
Flag that defines whether ownership type (like llc) should be cut-off.

Value

replace_customer_name() returns a character string.

get_replacements() returns a dplyr::tibble with columns from and to.

Examples

library(dplyr)

replace_customer_name("A. and B")

#> [1] "ab"

replace_customer_name("Acuity Brands Inc")

#> [1] "acuitybrands inc"

replace_customer_name(c("3M Company", "Abbott Laboratories", "AbbVie Inc."))

#> [1] "threem co"          "abbottlaboratories" "abbvie inc"        

custom_replacement <- tibble(from = "AAAA", to = "B")
replace_customer_name("Aa Aaaa", from_to = custom_replacement)

#> [1] "aab"

neutral_replacement <- tibble(from = character(0), to = character(0))
replace_customer_name("Company Name Owner", from_to = neutral_replacement)

#> [1] "companynameowner"

replace_customer_name(
  "Company Name Owner",
  from_to = neutral_replacement,
  ownership = "owner",
  remove_ownership = TRUE
)

#> [1] "companyname"

get_replacements()

#> # A tibble: 61 x 2
#>    from    to   
#>    <chr>   <chr>
#>  1 " and " " & "
#>  2 " en "  " & "
#>  3 " och " " & "
#>  4 " und " " & "
#>  5 (pjsc)  ""   
#>  6 (pte)   ""   
#>  7 (pvt)   ""   
#>  8 0       null 
#>  9 1       one  
#> 10 2       two  
#> # … with 51 more rows

append_replacements <- get_replacements() %>%
  add_row(
    .before = 1,
    from = c("AA", "BB"), to = c("alpha", "beta")
  )
append_replacements

#> # A tibble: 63 x 2
#>    from    to   
#>    <chr>   <chr>
#>  1 AA      alpha
#>  2 BB      beta 
#>  3 " and " " & "
#>  4 " en "  " & "
#>  5 " och " " & "
#>  6 " und " " & "
#>  7 (pjsc)  ""   
#>  8 (pte)   ""   
#>  9 (pvt)   ""   
#> 10 0       null 
#> # … with 53 more rows

# And in combination with `replace_customer_name()`
replace_customer_name(c("AA", "BB", "1"), from_to = append_replacements)

#> [1] "alpha" "beta"  "one"