Restructure loanbook dataset (lbk) in preparation for fuzzy matching

This function restructures a loanbook dataset (lbk) in preparation for the fuzzy matching process. Most notably, it outputs an the alias column from values in the columns name_direct_loantaker and name_ultimate_parent.

restructure_loanbook_for_matching(data, overwrite = NULL)

Arguments

data
A dataframe. Should be a loanbook.

overwrite
A dataframe used to overwrite the sector and/or name columns of a particular direct loantaker or ultimate parent. To overwrite only sector, the value in the name column should be NA.

Value

A dataframe with unique combinations of name + sector, including all IDs, and with elements already manually overwritten.

See also

Examples

library(r2dii.dataraw)

loanbook_demo %>%
  restructure_loanbook_for_matching()

#> Uniquifying `id_direct_loantaker` & `id_ultimate_parent`.FALSE

#> Adding new columns `sector` and `borderline`.

#> # A tibble: 638 x 6
#>    level       id    name                        sector source alias            
#>    <chr>       <chr> <chr>                       <chr>  <chr>  <chr>            
#>  1 direct_loa… C294  Yuamen Xinneng Thermal Pow… power  loanb… yuamenxinnengthe…
#>  2 ultimate_p… UP15  Alpine Knits India Pvt. Li… power  loanb… alpineknitsindia…
#>  3 direct_loa… C293  Yuamen Changyuan Hydropowe… power  loanb… yuamenchangyuanh…
#>  4 ultimate_p… UP84  Eco Wave Power              power  loanb… ecowavepower     
#>  5 direct_loa… C292  Yuama Ethanol Llc           power  loanb… yuamaethanol llc 
#>  6 ultimate_p… UP288 University Of Iowa          power  loanb… universityofiowa 
#>  7 direct_loa… C299  Yudaksel Holding A.S        power  loanb… yudakselhldgs as 
#>  8 ultimate_p… UP54  China Electric Power (Fuji… power  loanb… chinaelectricpow…
#>  9 direct_loa… C305  Yukon Energy Corp 1736      power  loanb… yukonenergycorpo…
#> 10 ultimate_p… UP104 Garland Power & Light       power  loanb… garlandpowerlight
#> # … with 628 more rows

loanbook_demo %>%
  restructure_loanbook_for_matching(overwrite = overwrite_demo)

#> Uniquifying `id_direct_loantaker` & `id_ultimate_parent`.FALSE

#> Adding new columns `sector` and `borderline`.

#> # A tibble: 638 x 6
#>    level       id    name                        sector source alias            
#>    <chr>       <chr> <chr>                       <chr>  <chr>  <chr>            
#>  1 direct_loa… C294  yuamen power company        coal   manual yuamenpower co   
#>  2 ultimate_p… UP15  alpine india                power  manual alpineindia      
#>  3 direct_loa… C293  Yuamen Changyuan Hydropowe… power  loanb… yuamenchangyuanh…
#>  4 ultimate_p… UP84  Eco Wave Power              power  loanb… ecowavepower     
#>  5 direct_loa… C292  Yuama Ethanol Llc           power  loanb… yuamaethanol llc 
#>  6 ultimate_p… UP288 University Of Iowa          power  loanb… universityofiowa 
#>  7 direct_loa… C299  Yudaksel Holding A.S        power  loanb… yudakselhldgs as 
#>  8 ultimate_p… UP54  China Electric Power (Fuji… power  loanb… chinaelectricpow…
#>  9 direct_loa… C305  Yukon Energy Corp 1736      power  loanb… yukonenergycorpo…
#> 10 ultimate_p… UP104 Garland Power & Light       power  loanb… garlandpowerlight
#> # … with 628 more rows