Score the similarity between two strings

This function computes pairwise string similarities between elements of character vectors a and b, where the vector with less elements is recycled. It is a thin wrapper around stringdist::stringdist() with defaults set to the most common parameters used in 2dii analyses.

score_string_similarity(a, b, ..., method = "jw", p = 0.1)

Arguments

a
R object (target); will be converted by as.character.

b
R object (source); will be converted by as.character.

...
Additional arguments are passed on to stringdist::stringsim.

method
Method for distance calculation. One of c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"). See stringdist::stringdist-metrics.

p
Penalty factor for Jaro-Winkler distance. The valid range for p is 0 <= p <= 0.25. If p=0 (default), the Jaro-distance is returned. Applies only to method='jw'.

Value

Returns a vector with similarities, which are values between 0 and 1 where 1 corresponds to perfect similarity (distance 0) and 0 to complete dissimilarity. NA is returned when stringdist returns NA. Distances equal to Inf are truncated to a similarity of 0.

Details

The similarity is calculated by first calculating the distance using stringdist, dividing the distance by the maximum possible distance, and substracting the result from 1. This results in a score between 0 and 1, with 1 corresponding to complete similarity and 0 to complete dissimilarity. Note that complete similarity only means equality for distances satisfying the identity property. This is not the case e.g. for q-gram based distances (for example if q=1, anagrams are completely similar). For distances where weights can be specified, the maximum distance is currently computed by assuming that all weights are equal to 1.

See also

Examples

# Clear extreemes
score_string_similarity("aa", "aa")

#> [1] 1

score_string_similarity("aa", "bb")

#> [1] 0

# Unclear extreemes
score_string_similarity("ab", "ba")

#> [1] 0

identical(
  score_string_similarity("ab", "ac"),
  score_string_similarity("ac", "ab")
)

#> [1] TRUE

score_string_similarity(c("fewer", "items", "get", "recycled"), "recycled")

#> [1] 0.5500000 0.4416667 0.4861111 1.0000000