Add the column sector (as classified by 2dii) and borderline

This function adds two columns, sector and borderline:

sector: sector gives climate-relevant sectors looked up from standard sector classifications in production databases.

borderline: Indicates if the classification is borderline or not.

bridge_sector(data)

Arguments

data
A loanbook dataframe.

Value

A loanbook dataframe with additional sector and borderline columns.

Details

A company acts in a climate-relevant sector if it meets two conditions:

It is classified in one of the sectors that translate 1 to 1 from a a standard sector classification to a climate-relevant sector covered by the PACTA analysis (e.g. utilities to power).

It is classified in a sector that can but does not necessarily map to a climate-relevant sector covered by the PACTA analysis (e.g. power grid/network to Power, general mining to coal mining).

For a list of sector classification systems see r2dii.dataraw::classification_bridge.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Must be attached
library(r2dii.dataraw)

#> Loading required package: r2dii.utils

out <- bridge_sector(loanbook_demo)
new_columns <- setdiff(names(out), names(loanbook_demo))

out %>%
  select(new_columns, everything())

#> # A tibble: 320 x 21
#>    sector borderline id_loan id_direct_loant… name_direct_loa… id_intermediate…
#>    <chr>  <lgl>      <chr>   <chr>            <chr>            <chr>           
#>  1 power  TRUE       L1      C294             Yuamen Xinneng … <NA>            
#>  2 power  TRUE       L2      C293             Yuamen Changyua… <NA>            
#>  3 power  TRUE       L3      C292             Yuama Ethanol L… IP5             
#>  4 power  TRUE       L4      C299             Yudaksel Holdin… <NA>            
#>  5 power  TRUE       L5      C305             Yukon Energy Co… <NA>            
#>  6 power  TRUE       L6      C304             Yukon Developme… <NA>            
#>  7 power  TRUE       L7      C227             Yaugoa-Zapadnay… <NA>            
#>  8 power  TRUE       L8      C303             Yueyang City Co… <NA>            
#>  9 power  TRUE       L9      C301             Yuedxiu Corp One IP10            
#> 10 power  TRUE       L10     C302             Yuexi County AA… <NA>            
#> # … with 310 more rows, and 15 more variables:
#> #   name_intermediate_parent_1 <chr>, id_ultimate_parent <chr>,
#> #   name_ultimate_parent <chr>, loan_size_outstanding <dbl>,
#> #   loan_size_outstanding_currency <chr>, loan_size_credit_limit <dbl>,
#> #   loan_size_credit_limit_currency <chr>, sector_classification_system <chr>,
#> #   sector_classification_input_type <chr>,
#> #   sector_classification_direct_loantaker <dbl>, fi_type <chr>,
#> #   flag_project_finance_loan <chr>, name_project <lgl>,
#> #   lei_direct_loantaker <lgl>, isin_direct_loantaker <lgl>