Match a loanbook (lbk) and asset-level datasets (ald) by the name_* columns

match_name() scores the match between names in a loanbook dataset (columns name_direct_loantaker and name_ultimate_parent) with names in an asset-level dataset (colum n name_company). The raw names are first transformed and stored in a simpler_name column, then the similarity between the simpler_name columns in each of the loanbook and ald datasets is scored using stringdist::stringsim().

match_name(
  loanbook,
  ald,
  by_sector = TRUE,
  min_score = 0.8,
  method = "jw",
  p = 0.1,
  overwrite = NULL
)

Arguments

loanbook
Dataframes with simpler_name and optionally sector columns.

ald
Dataframes with simpler_name and optionally sector columns.

by_sector
Should the combinations be done by sector?

min_score
A number (length-1) to set the minimum score values you want to pick.

method
Method for distance calculation. The default is "osa", see stringdist-metrics.

p
Penalty factor for Jaro-Winkler distance. The valid range for p is 0 <= p <= 0.25. If p=0 (default), the Jaro-distance is returned. Applies only to method='jw'.

overwrite
A dataframe used to overwrite the sector and/or name columns of a particular direct loantaker or ultimate parent. To overwrite only sector, the value in the name column should be NA.

Value

A dataframe with the same columns as the loanbook data with additional columns: id_lkb, sector_lbk, sector_ald, source_lbk, simpler_name_lbk, simpler_name_ald, score, name_ald.

Details

The process to create the simpler_name columns applies best practices commonly used in name matching algorithms, such as:

Remove special characters.

Replace language specific characters.

Abbreviate certain names to reduce their importance in the matching.

Spell out numbers to increase their importance.

See also

Examples

library(dplyr)
library(r2dii.dataraw)

match_name(loanbook_demo, ald_demo)

#> # A tibble: 450 x 27
#>    simpler_name_lbk simpler_name_ald score id_lbk sector_lbk source_lbk name_ald
#>    <chr>            <chr>            <dbl> <chr>  <chr>      <chr>      <chr>   
#>  1 astonmartin      astonmartin          1 UP23   automotive loanbook   aston m…
#>  2 avtozaz          avtozaz              1 UP25   automotive loanbook   avtozaz 
#>  3 bogdan           bogdan               1 UP36   automotive loanbook   bogdan  
#>  4 chauto           chauto               1 UP52   automotive loanbook   ch auto 
#>  5 chehejia         chehejia             1 UP53   automotive loanbook   chehejia
#>  6 chtcauto         chtcauto             1 UP58   automotive loanbook   chtc au…
#>  7 dongfenghonda    dongfenghonda        1 UP79   automotive loanbook   dongfen…
#>  8 dongfengluxgen   dongfengluxgen       1 UP80   automotive loanbook   dongfen…
#>  9 electricmobilit… electricmobilit…     1 UP89   automotive loanbook   electri…
#> 10 faradayfuture    faradayfuture        1 UP94   automotive loanbook   faraday…
#> # … with 440 more rows, and 20 more variables: sector_ald <chr>,
#> #   name_ultimate_parent_lbk <chr>, name_direct_loantaker_lbk <chr>,
#> #   id_loan_lbk <chr>, id_direct_loantaker_lbk <chr>,
#> #   id_intermediate_parent_1_lbk <chr>, name_intermediate_parent_1_lbk <chr>,
#> #   id_ultimate_parent_lbk <chr>, loan_size_outstanding_lbk <dbl>,
#> #   loan_size_outstanding_currency_lbk <chr>, loan_size_credit_limit_lbk <dbl>,
#> #   loan_size_credit_limit_currency_lbk <chr>,
#> #   sector_classification_system_lbk <chr>,
#> #   sector_classification_input_type_lbk <chr>,
#> #   sector_classification_direct_loantaker_lbk <dbl>, fi_type_lbk <chr>,
#> #   flag_project_finance_loan_lbk <chr>, name_project_lbk <lgl>,
#> #   lei_direct_loantaker_lbk <lgl>, isin_direct_loantaker_lbk <lgl>

match_name(
  loanbook_demo, ald_demo,
  min_score = 0.9,
  by_sector = FALSE
)

#> # A tibble: 616 x 27
#>    simpler_name_lbk simpler_name_ald score id_lbk sector_lbk source_lbk name_ald
#>    <chr>            <chr>            <dbl> <chr>  <chr>      <chr>      <chr>   
#>  1 abahydropowerge… abahydropowerge…     1 UP1    power      loanbook   aba hyd…
#>  2 achinskyglinozi… achinskyglinozi…     1 UP2    cement     loanbook   achinsk…
#>  3 affinityrenewab… affinityrenewab…     1 UP3    power      loanbook   affinit…
#>  4 africaoil corp   africaoil corp       1 C2     oil and g… loanbook   africa …
#>  5 africonshp sa    africonshp sa        1 UP4    shipping   loanbook   africon…
#>  6 agnisteelspriva… agnisteelspriva…     1 UP5    power      loanbook   agni st…
#>  7 agrenewables     agrenewables         1 UP6    power      loanbook   agrenew…
#>  8 airasiaxbhd      airasiaxbhd          1 C3     aviation   loanbook   airasia…
#>  9 airasiaxbhd      airasiaxbhd          1 UP7    aviation   loanbook   airasia…
#> 10 airbaltic        airbaltic            1 C4     aviation   loanbook   airbalt…
#> # … with 606 more rows, and 20 more variables: sector_ald <chr>,
#> #   name_ultimate_parent_lbk <chr>, name_direct_loantaker_lbk <chr>,
#> #   id_loan_lbk <chr>, id_direct_loantaker_lbk <chr>,
#> #   id_intermediate_parent_1_lbk <chr>, name_intermediate_parent_1_lbk <chr>,
#> #   id_ultimate_parent_lbk <chr>, loan_size_outstanding_lbk <dbl>,
#> #   loan_size_outstanding_currency_lbk <chr>, loan_size_credit_limit_lbk <dbl>,
#> #   loan_size_credit_limit_currency_lbk <chr>,
#> #   sector_classification_system_lbk <chr>,
#> #   sector_classification_input_type_lbk <chr>,
#> #   sector_classification_direct_loantaker_lbk <dbl>, fi_type_lbk <chr>,
#> #   flag_project_finance_loan_lbk <chr>, name_project_lbk <lgl>,
#> #   lei_direct_loantaker_lbk <lgl>, isin_direct_loantaker_lbk <lgl>