Match a loanbook and asset-level datasets (ald) by the name_* columns

match_name() scores the match between names in a loanbook dataset (columns can be name_direct_loantaker, name_intermediate_parent_*and name_ultimate_parent) with names in an asset-level dataset (column name_company). The raw names are first internally transformed then the similarity between transformed names in each of the loanbook and ald datasets is scored using stringdist::stringsim().

match_name(
  loanbook,
  ald,
  by_sector = TRUE,
  min_score = 0.8,
  method = "jw",
  p = 0.1,
  overwrite = NULL
)

Arguments

loanbook, ald
Dataframes structured like r2dii.dataraw::loanbook_demo and r2dii.dataraw::ald_demo.

by_sector
Should the combinations be done by sector?

min_score
A number (length-1) to set the minimum score values you want to pick.

method
Method for distance calculation. One of c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"). See stringdist::stringdist-metrics.

p
Penalty factor for Jaro-Winkler distance. The valid range for p is 0 <= p <= 0.25. If p=0 (default), the Jaro-distance is returned. Applies only to method='jw'.

overwrite
A dataframe used to overwrite the sector and/or name columns of a particular direct loantaker or ultimate parent. To overwrite only sector, the value in the name column should be NA and vice-versa. This file can be used to manually match entities to ald.

Value

A dataframe with the same groups (if any) and columns as loanbook, and the additional columns:

id_2dii - an id used internally by match_name() to distinguish entities

level - the level of granularity that the loan was matched at (e.g direct_loantaker or ultimate_parent)

sector - the sector of the loanbook company

sector_ald - the sector of the ald company

name - the name of the loanbook company

name_ald - the name of the ald company

score - the score of the match (manually set this to 1 prior to calling prioritize() to validate the match)

source - determines the source of the match. (equal to loanbook unless the match is from overwrite

The returned rows depend on the argument min_value and the result of the column score for each loan:

If any row has score equal to 1, match_name() returns all rows where score equals 1, dropping all other rows.

If no row has score equal to 1,match_name() returns all rows where score is equal to or greater than min_score. * If there is no match the output is a 0-row tibble with the expected column names -- for type stability.

Assigning aliases

The transformation process used to compare names between loanbook and ald datasets applies best practices commonly used in name matching algorithms:

Remove special characters.

Replace language specific characters.

Abbreviate certain names to reduce their importance in the matching.

Spell out numbers to increase their importance.

Handling grouped data

This function ignores but preserves existing groups them.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(r2dii.dataraw)

#> Loading required package: r2dii.utils

mini_loanbook <- sample_n(loanbook_demo, 10)

match_name(mini_loanbook, ald_demo)

#> # A tibble: 17 x 27
#>    id_loan id_direct_loant… name_direct_loa… id_intermediate… name_intermedia…
#>    <chr>   <chr>            <chr>            <chr>            <chr>           
#>  1 L173    C207             Toyota Motor Co… <NA>             <NA>            
#>  2 L197    C316             Zil              <NA>             <NA>            
#>  3 L173    C207             Toyota Motor Co… <NA>             <NA>            
#>  4 L197    C316             Zil              <NA>             <NA>            
#>  5 L201    C104             Inner Mongolia … <NA>             <NA>            
#>  6 L47     C257             Yolo County Flo… IP3              Yolo Regional D…
#>  7 L55     C249             Yiyang Baoyuan … <NA>             <NA>            
#>  8 L5      C305             Yukon Energy Co… <NA>             <NA>            
#>  9 L77     C146             Nanjing Guodian… <NA>             <NA>            
#> 10 L63     C241             Yingjiang Mengd… <NA>             <NA>            
#> 11 L77     C146             Nanjing Guodian… <NA>             <NA>            
#> 12 L63     C241             Yingjiang Mengd… <NA>             <NA>            
#> 13 L55     C249             Yiyang Baoyuan … <NA>             <NA>            
#> 14 L47     C257             Yolo County Flo… IP3              Yolo Regional D…
#> 15 L47     C257             Yolo County Flo… IP3              Yolo Regional D…
#> 16 L24     C282             Ypfb Corporacion <NA>             <NA>            
#> 17 L5      C305             Yukon Energy Co… <NA>             <NA>            
#> # … with 22 more variables: id_ultimate_parent <chr>,
#> #   name_ultimate_parent <chr>, loan_size_outstanding <dbl>,
#> #   loan_size_outstanding_currency <chr>, loan_size_credit_limit <dbl>,
#> #   loan_size_credit_limit_currency <chr>, sector_classification_system <chr>,
#> #   sector_classification_input_type <chr>,
#> #   sector_classification_direct_loantaker <dbl>, fi_type <chr>,
#> #   flag_project_finance_loan <chr>, name_project <lgl>,
#> #   lei_direct_loantaker <lgl>, isin_direct_loantaker <lgl>, id_2dii <chr>,
#> #   level <chr>, sector <chr>, sector_ald <chr>, name <chr>, name_ald <chr>,
#> #   score <dbl>, source <chr>

# May be a little slow
if (FALSE) {
match_name(
  mini_loanbook, ald_demo,
  min_score = 0.9,
  by_sector = FALSE
)
}